requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.9.1
click==8.1.7
cloudinary==1.34.0
coverage==7.3.2
Deprecated==1.2.14
dj-database-url==2.1.0
dj-rest-auth==5.0.1
Django==4.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
gunicorn==21.2.0
jwcrypto==1.5.0
oauthlib==3.2.2
pathspec==0.11.2
Pillow==10.0.1
psycopg2==2.9.8
pyasn1==0.5.0
pyenchant==3.2.2
PyJWT==2.8.0
python-jose==3.3.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
responses==0.23.3
rsa==4.9
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
types-PyYAML==6.0.12.12
urllib3==1.26.16
whitenoise==6.5.0
